backend/requirements.txt
fastapi
uvicorn
pydantic
supabase
python-dotenv
# Add: openai, etc. as needed
